test: Split git tests into test/git.spec.js
